fix: Small fixes
