+++ src/WingmanDataTrail/MetricsVariables/MetricsVariable.ts
@@ -1,10 +1,7 @@
 import { VariableHide, VariableRefresh, VariableSort } from '@grafana/data';
-import { QueryVariable, sceneGraph, type SceneObjectState } from '@grafana/scenes';
+import { QueryVariable, type SceneObjectState } from '@grafana/scenes';
 
 import { trailDS, VAR_FILTERS } from 'shared';
-import { NULL_GROUP_BY_VALUE } from 'WingmanDataTrail/Labels/LabelsDataSource';
-
-import { VAR_WINGMAN_GROUP_BY, type LabelsVariable } from '../Labels/LabelsVariable';
 
 export const VAR_METRICS_VARIABLE = 'metrics-wingman';
 
@@ -28,30 +25,5 @@ export class MetricsVariable extends QueryVariable {
       sort: VariableSort.alphabeticalAsc,
       hide: VariableHide.hideVariable,
     });
-  }
-
-  protected onActivate() {
-    // ensure that the correct metrics are loaded when landing: sometimes filters are not interpolated and fetching metric names gives all the results
-    const labelsVariable = sceneGraph.lookupVariable(VAR_WINGMAN_GROUP_BY, this) as LabelsVariable;
-
-    const updateQuery = (groupBy: string) => {
-      const matcher = groupBy !== NULL_GROUP_BY_VALUE ? `${groupBy}=~".+",$${VAR_FILTERS}` : `$${VAR_FILTERS}`;
-      const query = `label_values({${matcher}}, __name__)`;
-
-      if (query !== this.state.query) {
-        this.setState({ query });
-        this.refreshOptions();
-      }
-    };
-
-    updateQuery(labelsVariable.state.value as string);
-
-    this._subs.add(
-      labelsVariable.subscribeToState((newState, prevState) => {
-        if (newState.value !== prevState.value) {
-          updateQuery(newState.value as string);
-        }
-      })
-    );
   }
 }
